Conditional  casting and cleanup sampsharp-component
