+++ src/sampsharp-component/compat.hpp
@@ -1,4 +1,3 @@
-
 #pragma once
 
 #include <coreclr_delegates.h>
