+++ test/SampSharp.OpenMp.Entities.Commands.Tests/TestExtensions.cs
